agents: fix crash in HttpCurlGlobalInitializer
Make sure the HttpCurlGlobalInitializer static instance is created before
the OTLPAgent so it's destroyed after the OTLPAgent in __run_exit_handlers.
This fixes an insidious crash which happened if between the
HttpCurlGlobalInitializer and OTLPAgent destruction, a config update caused
the metrics_exporter_ to be reset to a specific exporter. This caused the
HttpCurlGlobalInitializer instance destructor to be called twice thus
crashing.

Fixes: https://github.com/nodesource/nsolid/issues/33
PR-URL: https://github.com/nodesource/nsolid/pull/34
Reviewed-by: Trevor Norris <trev.norris@gmail.com>