src: make nsolid::ThreadMetrics safer
The current ThreadMetrics API is not easy to use safely because the
`ThreadMetrics` instance needs to be alive while there is a pending
`Update()` callback to be called. Change the API to use shared_ptr, as
we define a new type `SharedThreadMetrics` as
`std::shared_ptr<ThreadMetrics>`.
This is a breaking change but considering previous API was completely
unsafe and apparently there're still no consumers I think is a good
change.

PR-URL: https://github.com/nodesource/nsolid/pull/37
Reviewed-by: Trevor Norris <trev.norris@gmail.com>