agents: add support for start/stopTrackingHeapObjects
This patch implements triggers for `startTrackingHeapObjects`
and `stopTrackingHeapObjects` in the heap profiler.

The implementation is based on what the inspector does:

The `startTrackingHeapObjects` will simply activate that in the
HeapProfiler and; the `stopTrackingHeapObjects` will create a heap
snapshot and deactivate the `TrackingHeapObjects` in the Heap
profiler.

This patch also includes the `trackAllocations` option supported.

Signed-off-by: Juan José Arboleda <soyjuanarbol@gmail.com>