nsolid: add start/stopTrackingHeapObjects APIs
This patch implements triggers for `startTrackingHeapObjects`
and `stopTrackingHeapObjects` in the heap profiler.

The implementation is based on what the inspector does:

The `startTrackingHeapObjects` will simply activate that in the
HeapProfiler and; the `stopTrackingHeapObjects` will create a heap
snapshot and deactivate the `TrackingHeapObjects` in the Heap
profiler.

This patch also includes the `trackAllocations` option supported.

PR-URL: https://github.com/nodesource/nsolid/pull/48
Signed-off-by: Juan José Arboleda <soyjuanarbol@gmail.com>
Co-authored-by: Santiago Gimeno <santiago.gimeno@gmail.com>
Reviewed-by: Trevor Norris <trev.norris@gmail.com>
Reviewed-by: Santiago Gimeno <santiago.gimeno@gmail.com>