tools: remove Report JS and format-cpp
The Report JS check is not necessary for us since we're downstream of
Node.js, but they also always fail. Remove them so it's no longer
checked.

We follow the cpp linter, but format-cpp adds additional checks that
make the code worse. Remove it and only rely on the linter.

PR-URL: https://github.com/nodesource/nsolid/pull/85
Co-authored-by: Santiago Gimeno <santiago.gimeno@gmail.com>