agents: fix StatsDTcp destruction
Add an `internal_state_` member variable which allows us to track
`uv_req_t` completions, so we can avoid to actually destroy the instance
before they are completed.
The API is changed so the destructor is private and to destroy the
instance a new public `close_and_delete()` method has been added.

PR-URL: https://github.com/nodesource/nsolid/pull/87
Reviewed-by: Trevor Norris <trev.norris@gmail.com>