agents: refactor profiling on ZmqAgent
In preparation of adding support for heap profiling and sampling,
extract the CPU Profiling logic into a 2 generic templated functions:
`start_prof()` and `got_prof()` which can be configured via specific
policies. Create the `CPUProfilePolicy` for now. Later on we will add
policies for heap profiling and sampling.
Encapsulate the state of every profile type into a `ProfileState`.

PR-URL: https://github.com/nodesource/nsolid/pull/90
Reviewed-by: Trevor Norris <trev.norris@gmail.com>