lib: add nsolid JS API for heap profiling
It involves two different set of API's.

The first one, is the classic `N|Solid` api which will create a profile
and send it to the Console. It consists of the following methods:
- `heapProfile()` and `heapProfileEnd()`.

The second allows to retrieve the heap profile via a `Readable` using
the following method:

```
const stream = heapProfileStream(threadId, duration, trackAllocations);
```

PR-URL: https://github.com/nodesource/nsolid/pull/101
Reviewed-by: Trevor Norris <trev.norris@gmail.com>