nsolid: support Heap Sampling
This patch adds support to the "sampled" heapsnapshot.

This new API methods in N|Solid allow users to sample allocations in the
V8 heap; internally it does the calls to the V8 Profiler for start
sampling and stop sampling.

This patch requires a new mutex and a new thread map to manage which
thread is currently sampling, same as CPU profiler and HeapSnapshot.

Signed-off-by: Juan José Arboleda <soyjuanarbol@gmail.com>