lib: add nsolid JS API for heap sampling
It involves two different set of API's.

The first one, is the classic `N|Solid` api which will create a profile
and send it to the Console. It consists of the following methods:
- `heapSampling()` and `heapSamplingEnd()`.

The second allows to retrieve the heap profile via a `Readable` using
the following method:

```
const stream = heapSamplingStream(threadId, duration, options);
```

PR-URL: https://github.com/nodesource/nsolid/pull/107
Reviewed-by: Trevor Norris <trev.norris@gmail.com>