agents: remove need for loop on create()
Not going to worry about the overhead of needing to call uv_loop_init()
on every call to create() since it's almost never called. This way we
don't need to pass in a uv_loop_t on the call.

If the overhead is a problem in the future then we'll change it so
there's a StatsDEndpoint loop just for that call and add a lock for
access so it's thread-safe.

PR-URL: https://github.com/nodesource/nsolid/pull/121
Reviewed-by: Santiago Gimeno <santiago.gimeno@gmail.com>