agents: have tcp/udp retry in statsd
Instead of handling the retry logic in the StatsDAgent, have each
connection handle its own endpoint. This way we can abstract the
connection class with a future virtual class and allow multiple
instances of StatsDAgent to be instantiated.

This change also involves other cleanup, and changing the way the timer
works so it retries to connect to the StatsD endpoint, after all IP
addresses have been tried, every 3 seconds.

PR-URL: https://github.com/nodesource/nsolid/pull/121
Reviewed-by: Santiago Gimeno <santiago.gimeno@gmail.com>