agents: add metrics transform API to agent
Allow users to register a callback that will transform the metrics into
a std::string that will be written to the endpoint.

In order to do this some of the internals needed to be rewritten so that
config() could be called from any thread. The internal usage of config()
has been renamed since it will always be called from the StatsDAgent
thread.

PR-URL: https://github.com/nodesource/nsolid/pull/123
Reviewed-by: Santiago Gimeno <santiago.gimeno@gmail.com>