deps: add grpc@1.52.0
The consumer should add the `grpc++` target (the C++ gRPC library) as a
dependency to add gRPC support.
A couple of other useful targets have been created, the most relevant for now
being `grpc_cpp_plugin` which creates the plugin to be used by `protoc` to
create the OTLP gRPC headers with the protobuf definitions.

Also, modify `test-strace-openat-openssl` so it takes into account an
additional `openat()` syscall.

PR-URL: https://github.com/nodesource/nsolid/pull/133
Reviewed-by: Trevor Norris <trev.norris@gmail.com>