agents: allow using OTLP env variable in OTLPAgent
The configuration of the OTLPAgent is IMHO a bit clunky. The are 2
configuration params/env variables: `NSOLID_OTLP` and
`NSOLID_OTLP_CONFIG`.
`NSOLID_OTLP` can take multiple values depending on the backend we're
planning to send the data to: `datadog`, `dynatrace`, `newrelic` and
`otlp`.
`NSOLID_OTLP_CONFIG` takes a JSON object with the config which has
different format depending on the `NSOLID_OTLP` value.
For the case of `otlp` the format of the config is something like this:
```
{
  url: 'http://endpoint:port',
  protocol: 'http' // or 'grpc'
}
```
This only gets us so far: what happens if I want to add a key in a
specific heep header? what if I want to send metrics to and endpoint and
traces to a different one? I could keep adding fields to the config
object but, do we want this?
Luckily the OTLP exporter spec defines a series of standard env
variables which allows for very granular configuration of the exporter
(https://opentelemetry.io/docs/specs/otel/protocol/exporter/).
These env variables are supported by default by `opentelemetry-cpp`
library we're using.
To enable this, a very slight modification of `NSOLID_OTLP` and
`NSOLID_OTLP_CONFIG` is needed. In the event we want the agent to use
these variables we'll need to configure only `NSOLID_OTLP=otlp` and not
configure `NSOLID_OTLP_CONFIG`, which was required until now. In case we
use this new configuration and no OTLP_* env variable is used, the agent
will be launched configure with the default values defined in
https://opentelemetry.io/docs/specs/otel/protocol/exporter/

So for example, now if we want to send metrics using grpc to endpoint1
and traces using http to endpoint2 I would do it this way:
```
OTEL_EXPORTER_OTLP_METRICS_PROTOCOL=grpc
OTEL_EXPORTER_OTLP_METRICS_ENDPOINT=http://endpoint1:${port}/v1/metrics
OTEL_EXPORTER_OTLP_TRACES_PROTOCOL=http/protobuf # this could be omitted
as it's the default
OTEL_EXPORTER_OTLP_TRACES_ENDPOINT=http://endpoint2:${port}/v1/metrics
./nsolid
```

PR-URL: https://github.com/nodesource/nsolid/pull/138
Reviewed-by: Trevor Norris <trev.norris@gmail.com>