agents: get telemetry data even if no cmd handle
We want the ZmqAgent to collect data (traces and/or metrics) even if the
ZMQ channels aren't fully up. This wasn't happening and we were losing
data because the tracing and metrics configuration options were not
handled until the ZmqCommandHandle wasn't ready.

Ideally we should be able to collect traces even on initial loop
iteration without the need of calling `nsolid.start()`.  This should go
into a follow-up PR though.

PR-URL: https://github.com/nodesource/nsolid/pull/140
Reviewed-by: Trevor Norris <trev.norris@gmail.com>