test: fix flaky test when run under load
Running the test with the following:

    $ ./tools/test.py -J --repeat=100 test/addons/nsolid-statsdagent/nsolid-statsdagent.js

caused the test to fail in various ways. The fix takes into account that
when the test is slowed down under load it's possible for the
StatsDAgent instance to be ready before it can be checked in JS, and
that the data might have been placed into a single call to 'data'.

Fixes: https://github.com/nodesource/nsolid/issues/142
PR-URL: https://github.com/nodesource/nsolid/pull/144
Reviewed-by: Santiago Gimeno <santiago.gimeno@gmail.com>