agents: implement SpanCollector helper class
SpanCollector is a class that allows to start collecting spans and
report them to a callback function running in a specific uv_loop_t
thread. The collector can be configured with a minimum span count and a
time interval to report the spans. In addition, the collector can be
configured with a transform function that will be applied to the spans
before calling the final callback.
Use this class to dry the code in the `ZmqAgent` and `OTLPAgent`.

PR-URL: https://github.com/nodesource/nsolid/pull/160
Reviewed-By: Trevor Norris <trev.norris@gmail.com>