agents: fix ZmqAgent asset JS API
When using the JS API to retrieve specific assets and send them to the
Console, there was some code paths where the data would be pushed to
the Zmq sockets from the JS thread instead of the agent thread. This is
a serious bug that could cause unexpected breakage.
Refactor the code to actually avoid the issue.
Also took the opportunity to expand the ProfileCollector class to also
support Heap Snapshots and use it in the agent.