src,agents: add support for source code collection
The runtime now stores the `scriptId` and `url` or `path` of all the
loaded cjs and esm modules.
A new grpc command has been added to retrieve the source code for any
`scriptId`-`url|path` pair by reading the source code from file.
In case the specific `scriptId`-`url|path` pair wasn't stored, it
returns an error.
In the rare case where a ESM is not loaded from a file URL, the code
would be stored on loading time.

PR-URL: https://github.com/nodesource/nsolid/pull/240
Reviewed-By: Rafael Gonzaga <rafael.nunu@hotmail.com>
PR-URL: https://github.com/nodesource/nsolid/pull/245